refactor: use smartPage as main entrypoint

+++ smart-page-spring/smart-page-spring-mongo/src/main/java/io/narsha/smartpage/spring/mongo/MongoQueryExecutor.java
@@ -66,7 +66,7 @@ public class MongoQueryExecutor implements QueryExecutor {
         filtered.stream()
             .map(map -> (T) convert(paginatedFilteredQuery.targetClass(), map, rowMapper))
             .toList();
-    long count = mongoTemplate.count(query.skip(-1).limit(-1), collection);
+    int count = Long.valueOf(mongoTemplate.count(query.skip(-1).limit(-1), collection)).intValue();
 
     return new SmartPageResult(res, count);
   }
